ModernWMS_PY/requirements.txt
# Web Framework
fastapi==0.110.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# Database
SQLAlchemy==2.0.31
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
PyJWT==2.9.0

# Data Validation
pydantic==2.8.2
pydantic-settings==2.2.1

# Utilities
python-dateutil==2.9.0.post0
python-slugify==8.0.4
email-validator==2.1.0.post1

# Development
pytest==8.1.1
pytest-cov==4.1.0
black==24.3.0
isort==5.13.2
mypy==1.8.0

# Documentation
fastapi-pagination==0.12.10

# Caching
redis==5.0.1
